Adding appropriate controllers and services and adding inability to logout when no user is logged in
