+++ src/main/java/com/fs/ecom/ecom_webapp/services/UserService.java
@@ -1,14 +1,20 @@
 package com.fs.ecom.ecom_webapp.services;
 
-import com.fs.ecom.ecom_webapp.dto.LoginDTO;
-import com.fs.ecom.ecom_webapp.dto.RegisterDTO;
-import com.fs.ecom.ecom_webapp.dto.UpdateDTO;
-import com.fs.ecom.ecom_webapp.dto.UserDetailsDTO;
+import com.fs.ecom.ecom_webapp.dto.*;
+import com.fs.ecom.ecom_webapp.exceptions.AddressNotFoundException;
+import com.fs.ecom.ecom_webapp.exceptions.UserNotFoundException;
+import com.fs.ecom.ecom_webapp.models.AddressBook;
 import com.fs.ecom.ecom_webapp.models.User;
+import com.fs.ecom.ecom_webapp.repositories.AddressRepository;
 import com.fs.ecom.ecom_webapp.repositories.UserRepository;
+import com.fs.ecom.ecom_webapp.security.service.JwtService;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -16,6 +22,12 @@ public class UserService {
 
     @Autowired
     UserRepository repository;
+
+    @Autowired
+    private AddressRepository addressRepository;
+
+    @Autowired
+    private JwtService jwtService;
 
     public User getUser (Long id) {
         return repository.getReferenceById(id);
@@ -25,7 +37,7 @@ public class UserService {
         repository.save(user);
     }
 
-    public UserDetailsDTO updateUser (UpdateDTO updateDTO) throws NullPointerException {
+    private UserDetailsDTO updateUser (UpdateDTO updateDTO) throws NullPointerException {
         User existingUser = repository.getReferenceById(updateDTO.getId());
         existingUser.setFirstName(updateDTO.getFirstName());
         existingUser.setLastName(updateDTO.getLastName());
@@ -45,12 +57,165 @@ public class UserService {
         repository.deleteAll();
     }
 
-    public UserDetailsDTO registerUser (RegisterDTO registerDTO) {
-        String firstName = registerDTO.getFirstName();
-        String lastName = registerDTO.getLastName();
-        User user = new User(firstName, lastName, registerDTO.defaultUserName(), registerDTO.getEmail(), registerDTO.getPassword(), registerDTO.getMobile());
-        User registeredUser = repository.save(user);
-        return new UserDetailsDTO(registeredUser);
+    public UserDetailsDTO loadUserByToken(String token) {
+        String subject = jwtService.extractSubject(token);
+
+        Optional<User> user = repository.findByEmail(subject);
+
+        UserDetailsDTO userDetailsDTO = new UserDetailsDTO();
+
+        try{
+            if (user.isPresent()) {
+                User temp = user.get();
+                userDetailsDTO.setId(temp.getId());
+                userDetailsDTO.setUserName(temp.getUserName());
+                userDetailsDTO.setMobile(temp.getMobile());
+                userDetailsDTO.setEmail(temp.getEmail());
+                userDetailsDTO.setFirstName(temp.getFirstName());
+                userDetailsDTO.setLastName(temp.getLastName());
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return userDetailsDTO;
+    }
+
+    public UserDetailsDTO updateUserProfile(UserDetailsDTO userDetailsDTO, String token) {
+        Optional<User> user = repository.findByEmail(jwtService.extractSubject(token));
+        UpdateDTO updateDTO = null;
+        if (user.isPresent()) {
+            updateDTO = new UpdateDTO(user.get());
+            updateDTO.setEmail(userDetailsDTO.getEmail());
+            updateDTO.setFirstName(userDetailsDTO.getFirstName());
+            updateDTO.setLastName(userDetailsDTO.getLastName());
+            updateDTO.setUserName(userDetailsDTO.getUserName());
+            updateDTO.setMobile(userDetailsDTO.getMobile());
+        }
+        return updateUser(updateDTO);
+    }
+
+    public String getJWTfromCookie(HttpServletRequest request) {
+        String token = null;
+        if (request.getCookies() != null) {
+            for (Cookie cookie : request.getCookies()) {
+                if ("JWT".equals(cookie.getName())) {
+                    token = cookie.getValue();
+                    return token;
+                }
+            }
+        }
+        return "No Token";
+    }
+
+    public User getUserFromRequest(HttpServletRequest request) throws UserNotFoundException {
+        String token = getJWTfromCookie(request);
+        String subject = jwtService.extractSubject(token);
+        Optional<User> user = repository.findByEmail(subject);
+        if(user.isPresent()){
+            return user.get();
+        }
+        else {
+            throw new UserNotFoundException("No User found from token");
+        }
+    }
+
+    public AddressDTO addAddress(AddressDTO addressDto, HttpServletRequest request) {
+        User user = null;
+        try {
+            user = getUserFromRequest(request);
+        } catch (UserNotFoundException e) {
+            e.printStackTrace();
+        }
+
+        boolean addressListIsEmpty = addressRepository.findByUserId(user.getId()).isEmpty();
+
+        if(addressDto.isDefault() && !addressListIsEmpty){
+            updateDefault(user);
+        } else if (addressListIsEmpty){
+            addressDto.setDefault(true);
+        }
+
+        AddressBook address = new AddressBook(addressDto);
+        address.setUser(user);
+
+        AddressBook savedAddress = addressRepository.save(address);
+
+        AddressDTO addressDto1 = new AddressDTO(savedAddress);
+        return addressDto1;
+    }
+
+    public List<AddressDTO> getAddresses(HttpServletRequest request) {
+        User user = null;
+        try {
+            user = getUserFromRequest(request);
+        } catch (UserNotFoundException e) {
+            e.printStackTrace();
+        }
+
+        List<AddressBook> addressBookList = addressRepository.findByUserId(user.getId());
+        List<AddressDTO> addresses = new ArrayList<>();
+        for(AddressBook address : addressBookList){
+            AddressDTO addressDto = new AddressDTO(address);
+            if(addressDto.isDefault()){
+                addresses.addFirst(addressDto);
+            }
+            else{
+                addresses.add(addressDto);
+            }
+        }
+
+        return addresses;
+    }
+
+    public void deleteAddress(AddressDTO addressDto, HttpServletRequest request) throws AddressNotFoundException {
+        User user = null;
+        try {
+            user = getUserFromRequest(request);
+        } catch (UserNotFoundException e) {
+            e.printStackTrace();
+        }
+        if(addressDto.getUserId() != user.getId()){
+            throw new AddressNotFoundException("Address not found");
+        }
+        else {
+            if(addressDto.isDefault()){
+                if(addressRepository.findByUserId(user.getId()).size()>1){
+                    setDefault(user);
+                }
+            }
+            addressRepository.deleteById(addressDto.getId());
+        }
+
+    }
+    public AddressDTO updateAddress(AddressDTO addressDto, HttpServletRequest request) throws AddressNotFoundException {
+        User user = null;
+        try {
+            user = getUserFromRequest(request);
+        } catch (UserNotFoundException e) {
+            e.printStackTrace();
+        }
+        if(addressDto.getUserId() != user.getId()){
+            throw new AddressNotFoundException("Address not found");
+        }
+        else {
+            if(addressDto.isDefault()){
+                updateDefault(user);
+            }
+            addressDto = new AddressDTO(addressRepository.save(new AddressBook(addressDto, user)));
+        }
+        return addressDto;
+    }
+
+    private void updateDefault(User user){
+        for(AddressBook address : user.getAddresses()){
+            if(address.isDefault()){
+                address.setDefault(false);
+            }
+        }
+    }
+
+    private void setDefault(User user){
+        addressRepository.findByUserId(user.getId()).getFirst().setDefault(true);
     }
 
 }
